moved GetStrainMeasure/GetStressMeasure from hpp to cpp
