+++ applications/GeoMechanicsApplication/custom_constitutive/small_strain_umat_3D_interface_law.cpp
@@ -10,7 +10,7 @@
 //  Main authors:    Vahid Galavi
 //
 
-#include "custom_constitutive/small_strain_umat_3D_interface_law.hpp"
+#include "custom_constitutive/small_strain_umat_3D_interface_law.h"
 #include "constitutive_law_dimension.h"
 
 namespace Kratos
@@ -139,6 +139,16 @@ void SmallStrainUMAT3DInterfaceLaw::SetValue(const Variable<Vector>& rVariable,
 SizeType SmallStrainUMAT3DInterfaceLaw::WorkingSpaceDimension() { return Dimension; }
 
 SizeType SmallStrainUMAT3DInterfaceLaw::GetStrainSize() const { return VoigtSize; }
+
+ConstitutiveLaw::StrainMeasure SmallStrainUMAT3DInterfaceLaw::GetStrainMeasure()
+{
+    return StrainMeasure_Infinitesimal;
+}
+
+ConstitutiveLaw::StressMeasure SmallStrainUMAT3DInterfaceLaw::GetStressMeasure()
+{
+    return StressMeasure_Cauchy;
+}
 
 std::string SmallStrainUMAT3DInterfaceLaw::Info() const { return "SmallStrainUMAT3DInterfaceLaw"; }
