+++ applications/GeoMechanicsApplication/custom_constitutive/small_strain_umat_law.h
@@ -159,13 +159,13 @@ public:
      * @brief Returns the expected strain measure of this constitutive law (by default Green-Lagrange)
      * @return the expected strain measure
      */
-    StrainMeasure GetStrainMeasure() override { return StrainMeasure_Infinitesimal; }
+    StrainMeasure GetStrainMeasure() override;
 
     /**
      * returns the stress measure of this constitutive law (by default 1st Piola-Kirchhoff stress in voigt notation)
      * @return the expected stress measure
      */
-    StressMeasure GetStressMeasure() override { return StressMeasure_Cauchy; }
+    StressMeasure GetStressMeasure() override;
 
     /**
      * @brief Computes the material response:
